+++ conda/gateways/disk/permissions.py
@@ -1,0 +1,94 @@
+# -*- coding: utf-8 -*-
+from __future__ import absolute_import, division, print_function, unicode_literals
+from logging import getLogger
+import errno
+import shutil
+import sys
+from conda import CondaError
+from errno import EACCES, EEXIST, ENOENT, EPERM
+from itertools import chain
+from logging import getLogger
+from os import (W_OK, access, chmod, getpid, link as os_link, listdir, lstat, makedirs, readlink,
+                rename, symlink, unlink, walk, stat)
+from os.path import abspath, basename, dirname, isdir, isfile, islink, join, lexists
+from shutil import rmtree
+from stat import S_IEXEC, S_IMODE, S_ISDIR, S_ISLNK, S_ISREG, S_IWRITE
+from time import sleep
+from uuid import uuid4
+
+from ...compat import lchmod, text_type
+from ...exceptions import CondaOSError
+from ...utils import on_win
+
+log = getLogger(__name__)
+
+
+def try_write(dir_path, heavy=False):
+    """Test write access to a directory.
+
+    Args:
+        dir_path (str): directory to test write access
+        heavy (bool): Actually create and delete a file, or do a faster os.access test.
+           https://docs.python.org/dev/library/os.html?highlight=xattr#os.access
+
+    Returns:
+        bool
+
+    """
+    if not isdir(dir_path):
+        return False
+    if on_win or heavy:
+        # try to create a file to see if `dir_path` is writable, see #2151
+        temp_filename = join(dir_path, '.conda-try-write-%d' % getpid())
+        try:
+            with open(temp_filename, mode='wb') as fo:
+                fo.write(b'This is a test file.\n')
+            backoff_unlink(temp_filename)
+            return True
+        except (IOError, OSError):
+            return False
+        finally:
+            backoff_unlink(temp_filename)
+    else:
+        return access(dir_path, W_OK)
+
+
+def make_writable(path):
+    try:
+        mode = lstat(path).st_mode
+        if S_ISDIR(mode):
+            chmod(path, S_IMODE(mode) | S_IWRITE | S_IEXEC)
+        elif S_ISREG(mode):
+            chmod(path, S_IMODE(mode) | S_IWRITE)
+        elif S_ISLNK(mode):
+            lchmod(path, S_IMODE(mode) | S_IWRITE)
+        else:
+            log.debug("path cannot be made writable: %s", path)
+    except Exception as e:
+        eno = getattr(e, 'errno', None)
+        if eno in (ENOENT,):
+            log.debug("tried to make writable, but didn't exist: %s", path)
+            raise
+        elif eno in (EACCES, EPERM):
+            log.debug("tried make writable but failed: %s\n%r", path, e)
+        else:
+            log.warn("Error making path writable: %s\n%r", path, e)
+            raise
+
+
+def recursive_make_writable(path):
+    # The need for this function was pointed out at
+    #   https://github.com/conda/conda/issues/3266#issuecomment-239241915
+    # Especially on windows, file removal will often fail because it is marked read-only
+    if isdir(path):
+        for root, dirs, files in walk(path):
+            for path in chain.from_iterable((files, dirs)):
+                try:
+                    exp_backoff_fn(make_writable, join(root, path))
+                except (IOError, OSError) as e:
+                    if e.errno == ENOENT:
+                        log.debug("no such file or directory: %s", path)
+                    else:
+                        raise
+    else:
+        exp_backoff_fn(make_writable, path)
